refactor(addons): extract dependency helpers into cli/addon_deps (#727)
The four depth-1 dependency helpers this PR added to addon_cmd (addon_self_ref,
resolve_declared_deps, install_declared_deps, refresh_pack_dependencies, plus
their test) pushed the file to 1623 lines; the LOC gate (scripts/loc-gate.sh,
limit 1500) forbids further growth of that file, so they move to a new
cli/addon_deps module with no behavior change.
